Rename package from command to cli across multiple files
